Add a level loader.
The new LevelLoader class can deserialize a JSON-encoded level format.
This format defines the level through RLE lists of tiles, taking up as
little information as possible. Each time in the list is an entry into a
master map of tiles, each of which can reside in its own spritesheet.