Fix issues with losing changes, and add a toggle for layer visiblity.
If a layer was modified and then the layer was switched, the changes
would be lost, since we'd reload from the serialized level data. Now
we load that only once, and use the in-memory scratch area for it.

There's also a toggle for changing the visibility to only show the
current layer, making it easy to see what's on which layer.