Calculate spritesheet coordinates and sizes automatically.
Another step toward better spritesheet management. Coordinates and sizes
are all computed automatically now. The caller only needs to say how
many rows/cols are in the spritesheet, and the sizes will be computed
from that. That includes the size of the Sprite object itself. This
makes it much easier to have dynamic image sizes.