Make it easier to stop movement after a collision.
The functions for adjusting the position after collision didn't work
very well, especially when dealing with custom collision rects. Now,
on_collision can return False to abort the movement.

The player has been updated with a custom collision for this. The head
is no longer collidable. That makes it much easier to move around.