Improve infected human approach AI.
The approach AI now only kicks in when the player is within 400 pixels
of the enemy.

The direction it chooses to look in is a bit more sane now. It will
figure out whether the distance is greater in the X or Y location, and
pick a direction based on that. Now they actually appear to walk toward
the player correctly.