Optimize the hell out of things.
This introduces some major optimizations that fix all the in-game
performance problems I've encountered with the new giant level.

Quad trees no longer return duplicate sprites. Previously they'd return
dups if a sprite existed in multiple quadrants. We now use a stamp on
each main iteration to ensure that we don't return more than one of each
sprite.

The single group for all sprites in a level has been removed. We were
iterating over that to do update() (which did nothing), and for calling
tick, and for debug rectangles. That group is gone. The layers now
handle the tick and drawing, and QuadTrees are used for the drawing.

Ticks are now opt-out. Tiles don't need ticks, so they clear a flag.
Instead of the aforementioned group, we now have a list of sprites on
each layer that want ticks, and only those are iterated through.

The events layer isn't loaded.