Add chunk loading.
Instead of loading and placing tiles for the entire level at once, we
now load the chunks as we need them. This applies only to the tiles, and
not to things programatically placed in the levels.

A chunk size is currently 10, 10 tiles, though this may change.

When drawing, if the camera region is covering new chunks, the new ones
are loaded in and any we don't need are unloaded. This happens
seamlessly behind the scenes.