Compress the chunk data for tiles.
The chunk data has been compressed by removing the maps and tile type
info and compressing things down into a tuple of normalized row, col, and
tile IDs. The tile IDs map back into a centralized list of filenames and
X/Y coords. This means we need only a minimal amount of information
in the chunks, and can quickly load it out.

This saves about 8MB for the world map.