Make the snake attack.
The snake can now attack. When within a certain range, it will strike,
heading straight toward the location where the player was standing when
the snake began the attack (and just past). If it collides with
anything, or reaches hte destination, it will stop and wait a few
seconds before thinking about attacking again.