Don't start sprites until their chunks load.
A sprite won't start until the chunk is loaded. This means that all the
mobs in the map will just be hanging out, chillin', until their chunks
are loaded.

Once a sprite is started, it stays started even if leaving the chunk.
The exception being map tiles, which are removed anyway.